refactor(test): move JSON tests to ServerModules subfolder
Move all JSON-related integration tests to a dedicated ServerModules
subfolder to mirror the source code organization:
- JsonCommandTests.cs
- JsonArrayCommandTests.cs
- JsonObjectCommandTests.cs
- JsonBatchTests.cs

Updated namespaces to Valkey.Glide.IntegrationTests.ServerModules.
