+++ tests/Valkey.Glide.IntegrationTests/ServerModules/JsonBatchTests.cs
@@ -1,9 +1,10 @@
 // Copyright Valkey GLIDE Project Contributors - SPDX Identifier: Apache-2.0
 
+using Valkey.Glide.IntegrationTests;
 using Valkey.Glide.Pipeline;
 using Valkey.Glide.ServerModules;
 
-namespace Valkey.Glide.IntegrationTests;
+namespace Valkey.Glide.IntegrationTests.ServerModules;
 
 /// <summary>
 /// Integration tests for JSON batch commands via the GlideJsonBatch static methods.
